multi-phase-sce/requirements.txt
matplotlib>=3.4
pyyaml==6.0
tensorflow-gpu==2.3.0
